requirements.txt
yfinance
pandas
ta
openai>=1.0
python-dotenv
requests
